requirements.txt
Flask==2.3.2
scipy==1.10.1
Flask-SQLAlchemy==3.0.3 
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
email-validator==0.5.0
flaskapp==1.0.5
python-json-logger==2.0.7
pytest==7.3.1
